% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_scores.R
\docType{data}
\name{auc_scores}
\alias{auc_scores}
\title{AUC scores of models}
\format{
a data frame with 1600 rows and 3 columns
}
\usage{
data(auc_scores)
}
\description{
Datasets \code{auc_scores} contains results of classification models,
}
\details{
\itemize{
\item model - name of model followed by a number indicating setting of hyperparameters
\item split - indicator of split
\item auc - area under the ROC curve
}
}
