% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embbeding_epp.R
\name{calculate_pca}
\alias{calculate_pca}
\title{PCA for models}
\usage{
calculate_pca(
  list_epp_score,
  transformation_formula = as.formula("DATASET~MODEL"),
  value.var = "EPP",
  ...
)
}
\arguments{
\item{list_epp_score}{list of data frames with epp scores, list of outputs from calculate_epp_score}

\item{transformation_formula}{formula passed to data.table::dcast function how transform data.
If we want PCA on models we use transformation_formula = as.formula("DATASET~MODEL"), otherwise e use transformation_formula = as.formula("MODEL~DATASET")}

\item{value.var}{colnames with epp score}

\item{...}{other arguments pass to PCA function}
}
\description{
PCA for models
}
