% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_epp.R
\name{calculate_probability}
\alias{calculate_probability}
\title{Calculate probability of winning}
\usage{
calculate_probability(epp_1, epp_2)
}
\arguments{
\item{epp_1}{EPP score of model 1.}

\item{epp_2}{EPP score of model 2.}
}
\description{
Probability of winning of one model 1 is the logit of the difference between EPP score of this model and opponent model 2.
}
