% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epp_ranking.R
\name{plot_epp_ranking}
\alias{plot_epp_ranking}
\title{Plot EPP ranking}
\usage{
plot_epp_ranking(epp, aggregation = mean, confidence_intervals = FALSE)
}
\arguments{
\item{epp}{object of class epp_results.}

\item{aggregation}{aggregation function.}

\item{confidence_intervals}{whether confidence intervals should be plotted. Only for "glmnet" estimation.}
}
\description{
Plot EPP ranking
}
