% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wins_ratio.R
\name{plot_wins_ratio}
\alias{plot_wins_ratio}
\title{Plot true ratio vs predicted probability}
\usage{
plot_wins_ratio(elo)
}
\arguments{
\item{elo}{object of class elo_results.}
}
\description{
Plot true ratio vs predicted probability
}
