% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_epp.R
\name{prepare_contrasts}
\alias{prepare_contrasts}
\title{Preparing matrix of contrasts}
\usage{
prepare_contrasts(actual_score)
}
\arguments{
\item{actual_score}{A data frame created with function calculate_actual_wins.}
}
\description{
Preparing matrix of contrasts
}
