% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{print.elo_test}
\alias{print.elo_test}
\title{Printing Summary of the Elo Test}
\usage{
\method{print}{elo_test}(x, ...)
}
\arguments{
\item{x}{elo_test object containing information about the test results.}

\item{...}{other parameters
is not specified then transformations for all variables are printed}
}
\value{
No return value, prints the structure of the object
}
\description{
Printing Summary of the Elo Test
}
