% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_elo.R
\name{print.epp_results}
\alias{print.epp_results}
\title{Printing EPP results}
\usage{
\method{print}{epp_results}(x, ...)
}
\arguments{
\item{x}{epp_results. The result of a function \code{\link{calculate_epp}}.}

\item{...}{other parameters}
}
\value{
No return value, prints the structure of the object
}
\description{
Printing EPP results
}
