% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_elo.R
\name{calculate_actual_wins}
\alias{calculate_actual_wins}
\title{Actual Results for Every Pair of Players}
\usage{
calculate_actual_wins(
  results,
  decreasing_metric = TRUE,
  compare_in_round,
  aggregate = TRUE
)
}
\arguments{
\item{results}{raw results}

\item{decreasing_metric}{Logical. If Score is decreasing metrics.}

\item{compare_in_round}{Logical. If Players should be compared only in the same Round. Setting to FALSE increase the number of Matches, but Score values of Players will be compared between different Rounds.}

\item{aggregate}{Logical. If results should be aggregated for every pair of Players. Otherwise, output will have many rows with binary response (according to amount of Rounds in the Tournament) for every pair of Players.}
}
\value{
data.frame
}
\description{
Calculate number of wins and loses for every pair of Players.
}
\details{
Naming convention, such as Player, Rounds, etc. comes from \href{https://arxiv.org/abs/2006.02293}{Gosiewska et al. (2020)}.
}
