% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_epp.R
\name{calculate_epp}
\alias{calculate_epp}
\title{Calulate EPP score for all models}
\usage{
calculate_epp(results, decreasing_metric = TRUE,
  compare_in_split = TRUE)
}
\arguments{
\item{results}{data frame with results for one dataset}

\item{decreasing_metric}{if used metric is decreasing. If TRUE Model with higher score value is better.}

\item{compare_in_split}{if compare models and parameters only in the same fold}
}
\description{
Calulate EPP score for all models
}
