% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_epp.R
\name{calculate_epp}
\alias{calculate_epp}
\title{Calulate EPP score for all models}
\usage{
calculate_epp(results, decreasing_metric = TRUE,
  compare_in_split = TRUE, keep_data = FALSE)
}
\arguments{
\item{results}{data frame with results for one dataset. Data should be in the following format.
First 3 columns should correspond to: model, split, score. See more in 'details' section.}

\item{decreasing_metric}{Logical. If TRUE used metric is considered as decreasing, that means a model with higher score value is considered as better model.
If FALSE used metric will be considered as increasing.}

\item{compare_in_split}{Logical. If TRUE compares models only in the same fold. If FALSE compares models across folds.}

\item{keep_data}{Logical. If FALSE only EPP scores will be returned. If TRUE original data frame with new 'epp' column will be returned.}
}
\description{
Calulate EPP score for all models
}
\details{
Format of the data frame passed via results parameter.
First column should correspond to a model. Dofferent settings of hyperparameters of the same model should have different values in this column.
Second column corresponds to indexes of splits. As EPP is based on Elo rating system, power of model is assessed by comparing its results
with other models on multiple data splits. Therefore, each model should be evaluated on multiple train-test splits.
Indexes of splits should be in this column. And should match across models.
Third column contains score used to evaluate models. It can be both decreasing or increasing metric,
just remember to set the \code{decreasing_metric} parameter accordingly.
The following columns can be of any kind.
}
\examples{
library(EloML)
data(auc_scores)
calculate_epp(auc_scores)

}
