% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epp_ranking.R
\name{plot_epp_ranking}
\alias{plot_epp_ranking}
\title{Plot EPP Meta-Score Ranking}
\usage{
plot_epp_ranking(
  epp,
  aggregation = mean,
  confidence_intervals = FALSE,
  show_player_names = FALSE
)
}
\arguments{
\item{epp}{epp_results. The result of a function \code{\link{calculate_epp}}.}

\item{aggregation}{Function. Aggregation function, for example mean or median.}

\item{confidence_intervals}{Logical. If confidence intervals should be plotted. Only for 'glm' estimation.}

\item{show_player_names}{Logical. If Players' names should be shown.}
}
\value{
gg
}
\description{
EPP Meta-Score values vs Aggreated values of Scores.
}
\details{
Naming convention, such as Player, Rounds, etc. comes from \href{https://arxiv.org/abs/2006.02293}{Gosiewska et al. (2020)}.
}
\examples{
library(EloML)
data(auc_scores)
epp <- calculate_epp(
  auc_scores[1:60, ],
  keep_columns=TRUE,
  keep_model = TRUE,
  estimation ="glm"
  )
plot_epp_ranking(epp, confidence_intervals = TRUE, show_player_names = TRUE)

}
