% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wins_ratio.R
\name{plot_wins_ratio}
\alias{plot_wins_ratio}
\title{Plot True Ratio vs EPP Probability}
\usage{
plot_wins_ratio(epp, random_sample = NULL, random_state = NULL)
}
\arguments{
\item{epp}{epp_results. The result of a function \code{\link{calculate_epp}}.}

\item{random_sample}{Numeric. If not NULL \verb{random sample} fraction of Players is sampled from the set of all Players.}

\item{random_state}{Numeric. Seed for sampling Players. Used only when \code{random_sample} is not NULL.}
}
\value{
gg
}
\description{
Plot of an empirical probability of winning versus probability of winning
estimated from EPP values.
}
\details{
Naming convention, such as Player, Rounds, etc. comes from \href{https://arxiv.org/abs/2006.02293}{Gosiewska et al. (2020)}.
}
\examples{
library(EloML)
data(auc_scores)
epp <- calculate_epp(
  auc_scores[1:100, ],
  keep_columns=TRUE,
  keep_model = TRUE,
  estimation ="glm"
  )
plot_wins_ratio(epp)

}
