% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{print.epp_test}
\alias{print.epp_test}
\title{Printing Summary of the EPP Test}
\usage{
\method{print}{epp_test}(x, ...)
}
\arguments{
\item{x}{epp_test. The result of a function \code{\link{test_players_diff}}.}

\item{...}{other parameters}
}
\value{
No return value, prints the structure of the object
}
\description{
Printing Summary of the EPP Test
}
