% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{test_players_diff}
\alias{test_players_diff}
\title{Test the differences between two EPP Scores}
\usage{
test_players_diff(epp, player1, player2)
}
\arguments{
\item{epp}{object of class epp_results.}

\item{player1}{Name of the first player to test}

\item{player2}{Name of the second player to test}
}
\description{
The test is based on the Wald test.
}
\examples{
data(auc_scores)
epp <- calculate_epp(auc_scores[1:400,], keep_model = TRUE, estimation = "glm")
test_players_diff(epp, "gbm_10", "gbm_2")

}
